LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/stylevalue-normalization/normalize-ident.tentative-expected.txt


FAIL CSS identifiers are normalized from String to CSSKeywordValues auto cannot be parsed as a width
FAIL CSS identifiers are normalized from CSSOM to CSSKeywordValues assert_equals: expected "CSSKeywordValue" but got "CSSStyleValue"
